Solve puzzle 01
Implements a function that contains the algorithim to calculate the fuel
requirement for number of mass. This is a supplied algorithim from the
challenge.

The data contained in input.data contains Module Mass values separated
by newlines.

The solution for the puzzle utilizes a reducer to take the module mass
data from input.data, calculates the fuel mass required for the module
mass, and then adds it to the accumulator