Solve puzzle 02
A module has a mass, that requires a certain amount of fuel. The fuel
for that module also has a mass that requires fuel.

Implementing a pair of functions that recursively calculate the fuel
required for the fuel required for a module. Then like puzzle 01,
utilize a reducer to sum up the fuel mass required.