Solve Day 02 - Part 1
Uses a recursive function to take the input data "intcode" and a cursor
(operating position of data). The operation to perform on each opcode
was provided specification. 1 use addition; 2 use multiplication; 99
end;

Various helper functions were implemented to try an clean up the work
that is being done in compute (main recursive function).

In day02.js there is an additional helper function that was added to
perform a replace-at-position operation. The puzzle instructions asked
to replace the values at 1 and 2, with 12 and 2.